Feature/#1 (#2)
* fix: #1 Create Pico_and_LED.h

Create Pico_and_LED.h and the code for it.

* fix: #1 Create Pico_and_LED.cpp

Create Pico_and_LED.cpp and the code for it.

* fix: #1 Create legend.txt

Create legend.txt and the content for it.

* fix: #1 Create main.cpp

Create main.cpp and the code for it. This is the demonstration file for the Pico_and_LED library.

* fix: #1 Create CMakeLists.txt

Create CMakeLists.txt and the content for it. This is for the demonstration file (main.cpp) to work with cmake.

* fix: #1 Added pico_sdk_import.cmake

Added pico_sdk_import.cmake for the demonstration file (main.cpp).

* fix: #1 Create Pico_and_LED_wiring_diagram

Create Pico_and_LED_wiring_diagram using Fritzing. This is for the demonstration file (main.cpp).