Feature/#4 (#5)
* fix: #4 Create Pico_and_Button.h

Create Pico_and_Button.h and the code for it.

* fix: #4 Create Pico_and_Button.cpp

Create Pico_and_Button.cpp and the code for it.

* fix: #4 Create legend.txt

Create legend.txt and the content for it.

* fix: #4 Create main.cpp

Create main.cpp and the code for it. This is the demonstration file for the Pico_and_Button library.

* fix: #4 Create CMakeLists.txt

Create CMakeLists.txt and the content for it. This is for the demonstration file (main.cpp) to work with cmake.

* fix: #4 Added pico_sdk_import.cmake

Added pico_sdk_import.cmake for the demonstration file (main.cpp).

* fix: #4 Create Pico_and_Button_wiring_diagram

Create Pico_and_Button_wiring_diagram using Fritzing. This is for the demonstration file (main.cpp).