Release/0.2 (#15)
* Feature/#1 (#2)

* fix: #1 Create Pico_and_LED.h

Create Pico_and_LED.h and the code for it.

* fix: #1 Create Pico_and_LED.cpp

Create Pico_and_LED.cpp and the code for it.

* fix: #1 Create legend.txt

Create legend.txt and the content for it.

* fix: #1 Create main.cpp

Create main.cpp and the code for it. This is the demonstration file for the Pico_and_LED library.

* fix: #1 Create CMakeLists.txt

Create CMakeLists.txt and the content for it. This is for the demonstration file (main.cpp) to work with cmake.

* fix: #1 Added pico_sdk_import.cmake

Added pico_sdk_import.cmake for the demonstration file (main.cpp).

* fix: #1 Create Pico_and_LED_wiring_diagram

Create Pico_and_LED_wiring_diagram using Fritzing. This is for the demonstration file (main.cpp).

* Feature/#4 (#5)

* fix: #4 Create Pico_and_Button.h

Create Pico_and_Button.h and the code for it.

* fix: #4 Create Pico_and_Button.cpp

Create Pico_and_Button.cpp and the code for it.

* fix: #4 Create legend.txt

Create legend.txt and the content for it.

* fix: #4 Create main.cpp

Create main.cpp and the code for it. This is the demonstration file for the Pico_and_Button library.

* fix: #4 Create CMakeLists.txt

Create CMakeLists.txt and the content for it. This is for the demonstration file (main.cpp) to work with cmake.

* fix: #4 Added pico_sdk_import.cmake

Added pico_sdk_import.cmake for the demonstration file (main.cpp).

* fix: #4 Create Pico_and_Button_wiring_diagram

Create Pico_and_Button_wiring_diagram using Fritzing. This is for the demonstration file (main.cpp).

* Feature/#6 (#7)

* fix: #6 Update Pico_and_LED.h

Update the code of Pico_and_LED.h.

* fix: #6 Update Pico_and_LED.cpp

Update the code of Pico_and_LED.cpp.

* fix: #6 Update main.cpp

Update the code of main.cpp.

* fix: #6 Update legend.txt

Update legend.txt and the content of it.

* Feature/#8 (#9)

* fix: #8 Update legend.txt (#9)
Update legend.txt and the content of it.

* Feature/#11 (#12)

* fix: #11 Update legend.txt (#12)  
Update legend.txt and the content of it.

* Feature/13 (#14)

* fix: #13 Update Pico_and_LED.h

Update the code of Pico_and_LED.h.

* fix: #13 Create Pico_and_Button.cpp 

Update the code of Pico_and_Button.cpp .

* fix: #13 Pico_and_Button.cpp

Update the code of Pico_and_Button.cpp.

* fix: #13 Update main.cpp.

Update the code of main.cpp.

* fix: #13 Update legend.txt

Update legend.txt and the content of it.