Feature/#16 (#17)
* fix: #16 Create Pico_and_Temperature.h

Create Pico_and_Temperature.h and the code for it.

* fix: #16 Create Pico_and_Temperature.cpp

Create Pico_and_Temperature.cpp and the code for it.

* fix: #16 Create Create legend.txt 

Create legend.txt and the content for it.

* fix: #16 Create main.cpp 

Create main.cpp and the code for it. This is the demonstration file for the Pico_and_Temperature library.

* fix: #16 Create CMakeLists.txt

Create CMakeLists.txt and the content for it. This is for the demonstration file (main.cpp) to work with cmake.

* fix: #16 Added pico_sdk_import.cmake

Added pico_sdk_import.cmake for the demonstration file (main.cpp).